+++ include/pv_general.h
@@ -7,11 +7,12 @@
 #define include_PV_GENERAL_H
 
 #include <stdbool.h>
+#include <stddef.h>
 
 /** @brief pointer arrayの内容数を返す
  * (実長さは番兵のNULL終端があるため、return+1)
  */
-int pv_general_get_parray_num(void **pointers);
+size_t pv_general_get_parray_num(void **pointers);
 
 
 bool pv_general_strtod(double *value, const char *str,
