+++ include/pv_io_util.h
@@ -15,7 +15,7 @@ typedef struct{
 
 double pv_io_util_get_double_from_str(const char *str);
 bool pv_io_util_get_pv_color_from_svg_str_rgba(PvColor *ret_color, const char *str);
-bool pv_read_args_from_str(double *args, int num_args, const char **str);
+bool pv_read_args_from_str(double *args, size_t num_args, const char **str);
 void pv_double_array_fill(double *dst, double value, int size);
 PvStrMap *pv_new_css_str_maps_from_str(const char *style_str);
 void pv_str_maps_free(PvStrMap *);
